gcp/generate-design-suggestion/requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.71.0
google-cloud-storage>=2.18.0
google-cloud-logging>=3.10.0
google-cloud-bigquery>=3.27.0
PyJWT[crypto]>=2.9.0
requests>=2.32.0
